WIP generators add to BloodLayer
